+++ include/libhal-arm-mcu/watchdog.hpp
@@ -4,27 +4,29 @@
 
 namespace hal {
 
-class watchdog {
-    /**
-    * @brief start the watchdog countdown
-    */
-    virtual void start_watchdog()=0;
-    /**
-    * @brief resets the watchdog countdown
-    */
-    virtual void reset_watchdog()=0;
-    /**
-    * @brief configures the watchdog to match countdown to specified time
-    */
-    virtual void set_watchdog_countdown_time(hal::time_duration wait_time)=0;
-    /**
-    * @brief checks if watchdog reset flag is set
-    */
-    virtual bool check_independent_watchdog_flag()=0;
-    /**
-    * @brief clears reset flags
-    */
-    virtual void clear_reset_flags()=0;
+class watchdog
+{
+public:
+  /**
+   * @brief start the watchdog countdown
+   */
+  virtual void start() = 0;
+  /**
+   * @brief resets the watchdog countdown
+   */
+  virtual void reset() = 0;
+  /**
+   * @brief configures the watchdog to match countdown to specified time
+   */
+  virtual void set_countdown_time(hal::time_duration wait_time) = 0;
+  /**
+   * @brief checks if watchdog reset flag is set
+   */
+  virtual bool check_flag() = 0;
+  /**
+   * @brief clears reset flags
+   */
+  virtual void clear_flag() = 0;
 };
 
 }  // namespace hal
